Fixed code

+++ app/src/main/java/ru/practicum/android/diploma/filter/data/dto/SettingsDto.kt
@@ -6,7 +6,7 @@ data class SettingsDto(
     val industry: IndustryDto,
     val country: CountryDto,
     val area: AreaDto,
-    val salary: Int,
+    val salary: Long,
     val onlyWithSalary: Boolean,
     val filterOn: Boolean
 )
